HACK: use filter-chain nodes if any, and hide them
This hack improves the policy to use filter-chain nodes by default if they are
available. Those nodes are then linked to the echo-cancel nodes if they are
available, otherwise the will be linked to the real audio devices.

One the other hand, similar to the previous hiding of echo-cancel nodes hack,
this patch also hides the filter chain nodes so that some applications don't get
confused when getting the real default device node.

This hack won't be needed once applications use the pipewire API to check what
real device the filter chain virtual nodes are using.