state-routes: only save route properties if it is available
Some ALSA devices seem to reset the properties of a route when the route becomes
unavailable (Eg when unplugging the headset from the Jack port). This fix avoids
saving the newly reset properties into the default-routes state file when that
happens. This makes sure the old properties are always restored when the route
becomes available again (Eg when plugging the headset back into the Jack port).